Implement game logic
Gui highlights correct small board that can be clicked.
Only correct small board can be accessed.
If the game on a small board is finished (X or O wins or there is
a draw), it displays a suitable mark and color.
If the small board is the board for the next move but the board is
inaccessible, then any other board can be chosen.